Fixing deltas of elements in a map or list
To allow MergeFrom to work, if the map or list type was a message, we were cloning it.  This clone can actually cause issues when normally adding one or many items to a map or list, as the original reference if it wasn't part of a parent message, can no longer be used to make changes.  Removing the clone for now.