Add support for subscribing to Message, List, and Map events (#28)
A typical pattern seen with event sourcing is a target will receive a stream of events and maintain its own replica of the state. A challenge that arises is how to update a view of that data without having to re-examine the entire data structure. These changes will allow for event listeners to be notified when changes are made to a specific Message, List or Map.