Removing message subscriptions
The ability to subscribe to changes is useful for clients that are receiving a stream of events.  There is a decent amount of overhead though even if this functionality is not used.  For now we are going to remove this functionality, until it can be added in a pluggable fashion.